Fix import logic to not import funcs from test files.
