+++ models/demos/llama3_subdevices/tt/model_config.py
@@ -33,11 +33,33 @@ from models.tt_transformers.tt.load_checkpoints import (
     standardize_hf_keys,
 )
 
-from tests.tt_eager.python_api_testing.unit_testing.misc.test_matmul_1d_gather_in0 import (
-    PREFETCHER_NOC1_GRID,
-)
 
-from tests.ttnn.unit_tests.operations.prefetcher_common import get_core_ranges
+PREFETCHER_NOC1_GRID = [
+    (6, 6),
+    (6, 7),
+    (6, 9),
+    (6, 0),
+    (6, 1),
+    (6, 2),
+    (6, 4),
+    (6, 5),
+    (5, 5),
+    (5, 6),
+    (5, 7),
+    (5, 9),
+    (5, 0),
+    (5, 1),
+    (5, 2),
+    (5, 4),
+    (1, 4),
+    (1, 5),
+    (1, 9),
+    (1, 0),
+    (2, 0),
+    (2, 4),
+    (2, 5),
+    (2, 9),
+]
 
 LM_HEAD_16_GRID = [
     (1, 0),
@@ -92,6 +114,220 @@ LM_HEAD_32_GRID = [
     (5, 0),
     (5, 1),
 ]
+
+
+def get_core_ranges(num_reader_cores, num_global_cb_receivers, is_functional_test):
+    """
+    Helper function to get all the relevant core ranges for dram prefetcher + matmul configuration
+    """
+
+    all_dram_cores = [ttnn.CoreCoord(idx, 0) for idx in range(12)]  # 12 DRAM banks
+
+    all_sender_cores = [
+        ttnn.CoreCoord(0, 9),
+        ttnn.CoreCoord(0, 0),
+        ttnn.CoreCoord(0, 4),
+        ttnn.CoreCoord(0, 5),
+        ttnn.CoreCoord(4, 0),
+        ttnn.CoreCoord(4, 9),
+        ttnn.CoreCoord(4, 1),
+        ttnn.CoreCoord(4, 7),
+        ttnn.CoreCoord(4, 6),
+        ttnn.CoreCoord(4, 2),
+        ttnn.CoreCoord(4, 4),
+        ttnn.CoreCoord(4, 5),
+    ]
+    dummy_sender_cores = [
+        ttnn.CoreCoord(0, 1),
+        ttnn.CoreCoord(0, 2),
+        ttnn.CoreCoord(0, 3),
+        ttnn.CoreCoord(0, 6),
+        ttnn.CoreCoord(0, 7),
+        ttnn.CoreCoord(0, 8),
+        ttnn.CoreCoord(4, 3),
+        ttnn.CoreCoord(4, 8),
+    ]
+
+    all_receiver_cores_list = [
+        (1, 9),
+        (2, 9),
+        (1, 0),
+        (2, 0),
+        (1, 4),
+        (2, 4),
+        (1, 5),
+        (2, 5),
+        (5, 0),
+        (6, 0),
+        (5, 9),
+        (6, 9),
+        (5, 1),
+        (6, 1),
+        (5, 7),
+        (6, 7),
+        (5, 6),
+        (6, 6),
+        (5, 2),
+        (6, 2),
+        (5, 4),
+        (6, 4),
+        (5, 5),
+        (6, 5),
+    ]
+
+    all_receiver_cores = [
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(*all_receiver_cores_list[idx]),
+                    ttnn.CoreCoord(*all_receiver_cores_list[idx + 1 if num_global_cb_receivers == 2 else idx]),
+                ),
+            ]
+        )
+        for idx in range(0, len(all_receiver_cores_list), 2)
+    ]
+
+    dummy_receiver_cores = [
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(3, 0),
+                    ttnn.CoreCoord(3, 0),
+                ),
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(1, 1),
+                    ttnn.CoreCoord(3, 1),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(1, 2),
+                    ttnn.CoreCoord(3, 2),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(1, 3),
+                    ttnn.CoreCoord(3, 3),
+                ),
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(3, 4),
+                    ttnn.CoreCoord(3, 4),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(3, 5),
+                    ttnn.CoreCoord(3, 5),
+                ),
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(1, 6),
+                    ttnn.CoreCoord(3, 6),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(1, 7),
+                    ttnn.CoreCoord(3, 7),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(1, 8),
+                    ttnn.CoreCoord(3, 8),
+                ),
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(3, 9),
+                    ttnn.CoreCoord(3, 9),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(5, 3),
+                    ttnn.CoreCoord(6, 3),
+                ),
+            ]
+        ),
+        ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(
+                    ttnn.CoreCoord(5, 8),
+                    ttnn.CoreCoord(6, 8),
+                ),
+            ]
+        ),
+    ]
+
+    dram_cores = all_dram_cores[:num_reader_cores]
+    hop_grid = []
+    mm_optimised_ring_cores = []
+    if not is_functional_test:
+        sender_cores = all_sender_cores
+        active_sender_cores = all_sender_cores[:num_reader_cores]
+        sender_cores.extend(dummy_sender_cores)
+        active_receiver_cores_list = all_receiver_cores_list[: num_reader_cores * num_global_cb_receivers]
+        receiver_cores = all_receiver_cores
+        receiver_cores.extend(dummy_receiver_cores)
+
+        worker_cores_range_set = ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(ttnn.CoreCoord(1, 0), ttnn.CoreCoord(3, 9)),
+                ttnn.CoreRange(ttnn.CoreCoord(5, 0), ttnn.CoreCoord(6, 9)),
+            ]
+        )
+
+        mm_optimised_ring_cores = PREFETCHER_NOC1_GRID
+        hop_grid = [
+            (3, 6),
+        ]
+    else:
+        sender_cores = [ttnn.CoreCoord(0, y) for y in range(num_reader_cores)]
+        active_sender_cores = sender_cores
+
+        active_receiver_cores_list = [
+            (x, y) for y in range(num_reader_cores) for x in range(1, num_global_cb_receivers + 1)
+        ]
+
+        receiver_cores = [
+            ttnn.CoreRangeSet(
+                [
+                    ttnn.CoreRange(
+                        ttnn.CoreCoord(*active_receiver_cores_list[idx * num_global_cb_receivers]),
+                        ttnn.CoreCoord(*active_receiver_cores_list[(idx + 1) * num_global_cb_receivers - 1]),
+                    )
+                ]
+            )
+            for idx in range(num_reader_cores)
+        ]
+
+        worker_cores_range_set = ttnn.CoreRangeSet(
+            [
+                ttnn.CoreRange(ttnn.CoreCoord(1, 0), ttnn.CoreCoord(2, 1)),
+            ]
+        )
+
+    return (
+        active_sender_cores,
+        dram_cores,
+        sender_cores,
+        active_receiver_cores_list,
+        receiver_cores,
+        worker_cores_range_set,
+        mm_optimised_ring_cores,
+        hop_grid,
+    )
 
 
 @dataclass
